refactor: переименованы scss файлы
