+++ src/components/ProjectPopup/InfoBlock/ProjectPopup.module.scss
