+++ src/components/ProjectPopup/PhotoBlock/PhotoBlock.module.scss
