+++ src/components/RoundArrowButton/RoundArrowButton.tsx
@@ -2,7 +2,7 @@ import cn from 'classnames';
 import React from 'react';
 
 import 'react-alice-carousel/lib/scss/alice-carousel.scss';
-import styles from './styles.module.scss';
+import styles from './RoundArrowButton.module.scss';
 
 import { ReactComponent as RoundArrowLeft } from '../../assets/icons/RoundArrowLeft.svg';
 import { ReactComponent as RoundArrowRight } from '../../assets/icons/RoundArrowRight.svg';
